+++ CHANGELOG.md
@@ -3,6 +3,7 @@ All notable changes to this project will be documented in this file.
 
 ## [next] - [now]
 
+* [Added] Missing directory handling (displays warning about missing directory and processes the other)
 * [Fixed] Option to set `false` for a `scripts` or `styles` path is now working properly
 * [Fixed] isProduction constant is now defined based on cli args, not only NODE_ENV